feat(sessions): Add confirmation prompt (#575)
Reopening of issue #572
Fixes issue #42
Translations added on PR 31 of revolt/translations

+++ src/pages/settings/panes/Sessions.tsx
@@ -20,6 +20,7 @@ import { useContext, useEffect, useState } from "preact/hooks";
 
 import { dayjs } from "../../../context/Locale";
 import { AppContext } from "../../../context/revoltjs/RevoltClient";
+import { useIntermediate } from "../../../context/intermediate/Intermediate";
 
 import Button from "../../../components/ui/Button";
 import Preloader from "../../../components/ui/Preloader";
@@ -38,6 +39,8 @@ export function Sessions() {
     );
     const [attemptingDelete, setDelete] = useState<string[]>([]);
     const history = useHistory();
+
+    const { openScreen } = useIntermediate();
 
     function switchPage(to: string) {
         history.replace(`/settings/${to}`);
@@ -212,21 +215,26 @@ export function Sessions() {
             <hr />
             <CategoryButton
                 onClick={async () => {
-                    // ! FIXME: add to rAuth
-                    const del: string[] = [];
-                    render.forEach((session) => {
-                        if (deviceId !== session._id) {
-                            del.push(session._id);
+                    openScreen({
+                        id: "sessions",
+                        confirm: async () => {
+                            // ! FIXME: add to rAuth
+                            const del: string[] = [];
+                            render.forEach((session) => {
+                                if (deviceId !== session._id) {
+                                    del.push(session._id);
+                                }
+                            });
+
+                            setDelete(del);
+
+                            for (const id of del) {
+                                await client.api.delete(`/auth/session/${id as ""}`);
+                            }
+
+                            setSessions(sessions.filter((x) => x._id === deviceId));
                         }
-                    });
-
-                    setDelete(del);
-
-                    for (const id of del) {
-                        await client.api.delete(`/auth/session/${id as ""}`);
-                    }
-
-                    setSessions(sessions.filter((x) => x._id === deviceId));
+                    })
                 }}
                 icon={<LogOut size={24} color={"var(--error)"} />}
                 action={"chevron"}
@@ -247,3 +255,4 @@ export function Sessions() {
         </div>
     );
 }
+
